pyproject.toml
[project]
name = "rag-mvp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.78",
    "chromadb>=1.5.5",
    "flashrank>=0.2.10",
    "google-generativeai>=0.8.6",
    "langchain>=1.2.13",
    "langchain-aws>=1.4.1",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-mcp-adapters>=0.2.2",
    "langgraph>=1.1.3",
    "llama-index>=0.14.19",
    "llama-index-embeddings-google-genai>=0.5.0",
    "llama-index-llms-google-genai>=0.9.1",
    "llama-index-vector-stores-chroma>=0.5.5",
    "llama-index-vector-stores-qdrant>=0.10.0",
    "llama-parse>=0.6.94",
    "opencv-python>=4.13.0.92",
    "pdf2image>=1.17.0",
    "pillow>=12.1.1",
    "pymupdf>=1.27.2.2",
    "pypdf>=6.9.2",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.1",
    "rich>=14.3.3",
    "streamlit>=1.55.0",
    "tenacity>=9.1.4",
    "tqdm>=4.67.3",
]
